+++ cli/cmd/aide-app/panel_other.go
@@ -1,0 +1,9 @@
+//go:build !darwin && cgo
+
+package main
+
+import "unsafe"
+
+func promotePanel(unsafe.Pointer) {}
+
+func showPanelNative(unsafe.Pointer) {}